Change filter behaviour at data extremities.
When filtering an image with a RutFilter FIR filter, any points
outside the bounds of the image that were required during convolution
were taken as zeros.  In order to reduce edge effects, this patch
alters the behaviour to use the value of the closest point on the edge
of the image.  This provides a substantial improvement.