Fix segfault bugs in ridgetool line stitching pass #2.
The pointer arithmetic that was used for converting the line sets into
lists was a little too clever for its own good, and would sometimes in
certain circumstances wander out of mapped memory and make ridgetool
explode.

This rewrites the process to work in image coordinates instead of with
pointers, and is more robust overall.