Build ridges with canonical ordering.
Make sure that ridge line conversion from set form to list form is
always carried out in such a way that the start of the line is
closer than the end of the line to the start of the image.  This
means that when the ridge line data is exported to an RDG file, it
is ordered by image position of line start.

Previously, there was a problem with indeterminate RDG file
ordering when multiple processes were used.  If a line spanned two
process's work areas, its position in the RDG file would depend on
which process got to create the root node. This caused problems
using RIO_KEY_IMAGE_CLASSIFICATION data to compare classifications,
because the ordering of the classification data was not guaranteed
to be consistent between RDG files generated in successive multi-
process ridgetool runs.